代码调整

+++ conv/conv3d_backprop_filter_v2/op_kernel/arch35/conv3d_backprop_filter_v2/conv3d_dw_v2_basic_block.h
@@ -228,7 +228,7 @@ protected:
         uint64_t maxNIter = Ceil(this->nCnt_ > 1 ? this->singleShapeN_ : nCoreTailAlign, this->dw_.ctx.tiling_->baseN);
 
         uint32_t barrierTriggerCnt = 0;
-        constexpr int8_t barrierThreshold = 14;
+        constexpr uint32_t barrierThreshold = 14;
         for (uint64_t i = 0; i < maxMIter; i++) {
             for (uint64_t j = 0; j < maxNIter; j++) {
                 bool isCurIter = (i < this->dw_.ctx.mIter_) && (j < this->dw_.ctx.nIter_);
